Crate diem_proptest_helpers

Structs

GrowingSubset

A set of elements, each with an associated key, that grows over time.

Index

Wrapper for proptest’s [Index][proptest::sample::Index] that allows AsRef to work.

RepeatVec

An efficient representation of a vector with repeated elements inserted.

ValueGenerator

Context for generating single values out of strategies.

Functions

pick_idxs

Given a maximum value max and a list of Index instances, picks integers in the range [0, max) uniformly randomly and without duplication.

pick_slice_idxs

Given a maximum value max and a slice of Index instances, picks integers in the range [0, max) uniformly randomly and without duplication.

with_stack_size

Creates a new thread with a larger stack size.