diem_proptest_helpers

Struct ValueGenerator

pub struct ValueGenerator { /* private fields */ }

Context for generating single values out of strategies.

Proptest is designed to be built around “value trees”, which represent a spectrum from complex values to simpler ones. But in some contexts, like benchmarking or generating corpuses, one just wants a single value. This is a convenience struct for that.

Implementations

impl ValueGenerator

pub fn new() -> Self

Creates a new value generator with the default RNG.

pub fn new_with_rng(rng: TestRng) -> Self

Creates a new value generator with provided RNG

pub fn deterministic() -> Self

Creates a new value generator with a deterministic RNG.

This generator has a hardcoded seed, so its results are predictable across test runs. However, a new proptest version may change the seed.

pub fn generate<S: Strategy>(&mut self, strategy: S) -> S::Value

Generates a single value for this strategy.

Panics if generating the new value fails. The only situation in which this can happen is if generating the value causes too many internal rejects.